Fixed two errors. Messages should work now!
I saw the Post_Message coding mistake before... Tidor (or whoever
originally wrote this) wasn't resetting the bcc-list to emptystring
after hitting the ceiling of bcc... we've never had problems until this
year b/c we had less than 80 on the list... I fixed it earlier this year
but forgot... and restoring the backup made it come back. yeah.